+++ include/steam/problem/WeightedLeastSqCostTerm.hpp
@@ -35,7 +35,7 @@ public:
   //////////////////////////////////////////////////////////////////////////////////////////////
   WeightedLeastSqCostTerm(const typename ErrorEvaluator<MEAS_DIM,MAX_STATE_SIZE>::ConstPtr& errorFunction,
            const typename NoiseModel<MEAS_DIM>::ConstPtr& noiseModel,
-           const LossFunction::ConstPtr& lossFunc);
+           const LossFunctionBase::ConstPtr& lossFunc);
 
   //////////////////////////////////////////////////////////////////////////////////////////////
   /// \brief Evaluate the cost of this term. Error is first whitened by the noise model
@@ -87,7 +87,7 @@ private:
   //////////////////////////////////////////////////////////////////////////////////////////////
   /// \brief Loss function
   //////////////////////////////////////////////////////////////////////////////////////////////
-  LossFunction::ConstPtr lossFunc_;
+  LossFunctionBase::ConstPtr lossFunc_;
 };
 
 //////////////////////////////////////////////////////////////////////////////////////////////
