+++ include/steam/problem/lossfunc/DcsLossFunc.hpp
@@ -1,0 +1,53 @@
+//////////////////////////////////////////////////////////////////////////////////////////////
+/// \file DcsLossFunc.hpp
+///
+/// \author Sean Anderson, ASRL
+//////////////////////////////////////////////////////////////////////////////////////////////
+
+#ifndef STEAM_DCS_LOSS_FUNCTION_HPP
+#define STEAM_DCS_LOSS_FUNCTION_HPP
+
+#include <Eigen/Core>
+#include <boost/shared_ptr.hpp>
+
+#include <steam/problem/lossfunc/LossFunctionBase.hpp>
+
+namespace steam {
+
+//////////////////////////////////////////////////////////////////////////////////////////////
+/// \brief Huber loss function class
+//////////////////////////////////////////////////////////////////////////////////////////////
+class DcsLossFunc : public LossFunctionBase
+{
+ public:
+
+  /// Convenience typedefs
+  typedef boost::shared_ptr<DcsLossFunc> Ptr;
+  typedef boost::shared_ptr<const DcsLossFunc> ConstPtr;
+
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  /// \brief Constructor -- k is the `threshold' based on number of std devs (1-3 is typical)
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  DcsLossFunc(double k) : k2_(k*k) {}
+
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  /// \brief Cost function (basic evaluation of the loss function)
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  virtual double cost(double whitened_error_norm) const;
+
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  /// \brief Weight for iteratively reweighted least-squares (influence function div. by error)
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  virtual double weight(double whitened_error_norm) const;
+
+ private:
+
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  /// \brief Huber constant
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  double k2_;
+};
+
+} // steam
+
+#endif // STEAM_DCS_LOSS_FUNCTION_HPP
