+++ include/steam/problem/lossfunc/LossFunctionBase.hpp
@@ -1,0 +1,47 @@
+//////////////////////////////////////////////////////////////////////////////////////////////
+/// \file LossFunctionBase.hpp
+///
+/// \author Sean Anderson, ASRL
+//////////////////////////////////////////////////////////////////////////////////////////////
+
+#ifndef STEAM_LOSS_FUNCTION_BASE_HPP
+#define STEAM_LOSS_FUNCTION_BASE_HPP
+
+#include <Eigen/Core>
+#include <boost/shared_ptr.hpp>
+
+namespace steam {
+
+//////////////////////////////////////////////////////////////////////////////////////////////
+/// \brief Base loss function class. A loss function must implement both the cost and
+///        weight functions. For example, the basic least-square L2 loss function has:
+///        cost: e^2, and weight: 1.
+//////////////////////////////////////////////////////////////////////////////////////////////
+class LossFunctionBase
+{
+public:
+
+  /// Convenience typedefs
+  typedef boost::shared_ptr<LossFunctionBase> Ptr;
+  typedef boost::shared_ptr<const LossFunctionBase> ConstPtr;
+
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  /// \brief Constructor
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  LossFunctionBase() {}
+
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  /// \brief Cost function (basic evaluation of the loss function)
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  virtual double cost(double whitened_error_norm) const = 0;
+
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  /// \brief Weight for iteratively reweighted least-squares (influence function div. by error)
+  //////////////////////////////////////////////////////////////////////////////////////////////
+  virtual double weight(double whitened_error_norm) const = 0;
+
+};
+
+} // steam
+
+#endif // STEAM_LOSS_FUNCTION_BASE_HPP
