+++ src/problem/lossfunc/DcsLossFunc.cpp
@@ -1,0 +1,36 @@
+//////////////////////////////////////////////////////////////////////////////////////////////
+/// \file DcsLossFunc.cpp
+///
+/// \author Sean Anderson, ASRL
+//////////////////////////////////////////////////////////////////////////////////////////////
+
+#include <steam/problem/lossfunc/DcsLossFunc.hpp>
+
+namespace steam {
+
+//////////////////////////////////////////////////////////////////////////////////////////////
+/// \brief Cost function (basic evaluation of the loss function)
+//////////////////////////////////////////////////////////////////////////////////////////////
+double DcsLossFunc::cost(double whitened_error_norm) const {
+  double e2 = whitened_error_norm*whitened_error_norm;
+  if (e2 <= k2_) {
+    return 0.5*e2;
+  } else {
+    return 2.0*k2_*e2/(k2_+e2) - 0.5*k2_;
+  }
+}
+
+//////////////////////////////////////////////////////////////////////////////////////////////
+/// \brief Weight for iteratively reweighted least-squares (influence function div. by error)
+//////////////////////////////////////////////////////////////////////////////////////////////
+double DcsLossFunc::weight(double whitened_error_norm) const {
+  double e2 = whitened_error_norm*whitened_error_norm;
+  if (e2 <= k2_) {
+    return 1.0;
+  } else {
+    double k2e2 = k2_ + e2;
+    return 4.0*k2_*k2_/(k2e2*k2e2);
+  }
+}
+
+} // steam
